Make MultiServerCluster abstract and prefix name with "Abstract"
Add concrete subclasses StableMultiServerCluster and DnsMultiServerCluster.
In the next commit StableMultiServerCluster will be renamed to simply
MultiServerCluster, but it has to be in a separate commit so as to not
lose the version history on what is now AbstractMultiServerCluster.

JAVA-2927
