+++ driver-core/src/test/unit/com/mongodb/internal/connection/AbstractServerDiscoveryAndMonitoringTest.java
@@ -147,7 +147,7 @@ public class AbstractServerDiscoveryAndMonitoringTest {
         if (settings.getMode() == ClusterConnectionMode.SINGLE) {
             cluster = new SingleServerCluster(clusterId, clusterSettings, factory);
         } else {
-            cluster = new MultiServerCluster(clusterId, clusterSettings, factory, null /* TODO */);
+            cluster = new StableMultiServerCluster(clusterId, clusterSettings, factory);
         }
     }
