+++ driver-core/src/test/unit/com/mongodb/internal/connection/ServerDiscoveryAndMonitoringMonitoringTest.java
@@ -102,7 +102,7 @@ public class ServerDiscoveryAndMonitoringMonitoringTest extends AbstractServerDi
                 if (newDescription.getString("topologyType").getValue().equals("Single")) {
                     assertEquals(SingleServerCluster.class, getCluster().getClass());
                 } else {
-                    assertEquals(MultiServerCluster.class, getCluster().getClass());
+                    assertEquals(StableMultiServerCluster.class, getCluster().getClass());
                 }
 
             } else if (eventDocument.containsKey("server_opening_event")) {
